# Conflicts:
#	src/main/java/com/example/chat/groupchatbackend/controllers/MessagesController.java
#	src/main/java/com/example/chat/groupchatbackend/repositories/ConversationsRepository.java

+++ src/main/java/com/example/chat/groupchatbackend/controllers/MessagesController.java
@@ -1,5 +1,8 @@
 package com.example.chat.groupchatbackend.controllers;
 
+import com.example.chat.groupchatbackend.Conversation;
+import com.example.chat.groupchatbackend.Message;
+import com.example.chat.groupchatbackend.UserContext;
 import com.example.chat.groupchatbackend.*;
 import com.example.chat.groupchatbackend.repositories.ConversationsRepository;
 import com.example.chat.groupchatbackend.repositories.MessagesRepository;
@@ -13,7 +16,6 @@ import javax.transaction.Transactional;
 import java.time.Instant;
 import java.time.LocalDateTime;
 import java.time.ZoneId;
-import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
 import java.util.Optional;
@@ -36,12 +38,22 @@ public class MessagesController {
     public List<Message> getAllMessages(@PathVariable String conversationName, @RequestParam(required = false) Long timestamp) {
         Conversation conversation = getConversation(conversationName);
 
-        if (timestamp == null) {
-            return getMessagesByDate(conversation, LocalDateTime.MIN);
-        } else {
-            LocalDateTime date = Instant.ofEpochMilli(timestamp).atZone(ZoneId.systemDefault()).toLocalDateTime();
-            return getMessagesByDate(conversation, date);
+    @GetMapping("/messages/channel/{id}")
+    public ResponseEntity getAllMessages(@PathVariable int id , @RequestParam(required = false) Long timestamp) {
+        Conversation conversation = conversationsRepository.findById(id).orElseThrow(() -> new RuntimeException("conversation doesn't exist"));
+        if (conversation.getConversationType().equals(ConversationType.CHANNEL)) {
+            LocalDateTime date = LocalDateTime.MIN;
+            if (timestamp != null) {
+                date = Instant.ofEpochMilli(timestamp).atZone(ZoneId.systemDefault()).toLocalDateTime();
+            }
+            List<Message> result =  getMessagesByDate(conversation, date);
+            return ResponseEntity
+                    .status(HttpStatus.OK)
+                    .body(result);
         }
+        return ResponseEntity
+                .status(HttpStatus.BAD_REQUEST)
+                .body("not channel");
     }
 
     private List<Message> getMessagesByDate(Conversation conversation, LocalDateTime date) {
@@ -121,19 +133,48 @@ public class MessagesController {
     @Transactional
     public Message editMessage(@PathVariable int id, @RequestBody String text) {
         Message message = getMessageById(id);
-        message.setText(text);
-        return message;
+        if (userContext.getCurrentUser().getId().equals(message.getSenderId())) {
+            message.setText(text);
+            return message;
+        } else {
+            throw new RuntimeException("You can't edit message");
+        }
     }
 
     @DeleteMapping("/messages/{id}")
     public ResponseEntity deleteMessage(@PathVariable int id) {
         Message message = getMessageById(id);
-        messagesRepository.delete(message);
-        return new ResponseEntity(HttpStatus.OK);
+        Conversation conversation = conversationsRepository.findByMessageId(id);
+        conversation.getMessages().removeIf(message1 -> message1.getId() == id);
+        conversationsRepository.save(conversation);
+        if (userContext.getCurrentUser().getId().equals(message.getSenderId())) {
+            messagesRepository.delete(message);
+            return new ResponseEntity(HttpStatus.OK);
+        } else {
+            return new ResponseEntity("You are not sender of message", HttpStatus.BAD_REQUEST);
+        }
     }
 
     private Message getMessageById(int messageId) {
         return messagesRepository.findById(messageId).orElseThrow(() -> new RuntimeException("message doesn't exist"));
+    }
+
+    @GetMapping("/messages/private/{conversationId}")
+    public List <Message> getPrivateConversationMessagesWithUser(@PathVariable int conversationId, @RequestParam(required = false) Long timestamp) {
+       Conversation conversation = conversationsRepository.findById(conversationId).
+               orElseThrow(() -> new RuntimeException("conversation doesn't exist"));
+
+       User user = userContext.getCurrentUser();
+       if(conversation.checkUserPresenceInConversation(user) && conversation.getConversationType().equals(ConversationType.DIRECT_MESSAGE)){
+           if (timestamp == null) {
+               return getMessagesByDate(conversation, LocalDateTime.MIN);
+           } else {
+               LocalDateTime date = Instant.ofEpochMilli(timestamp).atZone(ZoneId.systemDefault()).toLocalDateTime();
+               return getMessagesByDate(conversation, date);
+           }
+       } else {
+           throw new RuntimeException("User is not present in this conversation");
+       }
     }
 
 }
